PYTHON/pyproject.toml
[project]
name = "cortexa"
version = "1.1.2"
description = "Python implementation of Git-Context-Controller (GCC)"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Swadhin Biswas", email = "swadhinbiswas.cse@gmail.com" },
]
keywords = ["gcc", "git-context-controller", "llm", "agent", "memory", "context"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/swadhinbiswas/Cortexa"
Repository = "https://github.com/swadhinbiswas/Cortexa"
Issues = "https://github.com/swadhinbiswas/Cortexa/issues"

[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cortexa"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test.py", "test_*.py", "*_test.py"]
